+++ tests/test_loaders.py
@@ -14,11 +14,7 @@ CustomAbcStore = Mock()
 
 @pytest.fixture
 def get_crawler():
-    def _crawler(extended_settings={}):
-        settings = {
-            "EXTENSIONS": {"spider_feeder.loaders.StartUrlsLoader": 500},
-        }
-        settings.update(extended_settings)
+    def _crawler(settings={}):
         crawler = Crawler(Spider, settings=settings)
         crawler.spider = Spider("dummy")
         return crawler
@@ -42,38 +38,40 @@ def test_start_urls_loader_open_store_given_scheme(get_crawler, mocker, scheme, 
     mock = mocker.patch(store_cls)
     mock().__iter__.return_value = iter([('https://url1.com', {}), ('https://url2.com', {})])
 
-    crawler = get_crawler({'SPIDERFEEDER_INPUT_URI': f'{scheme}input_file.txt'})
-    StartUrlsLoader.from_crawler(crawler)
+    crawler = get_crawler({
+        'EXTENSIONS': {'spider_feeder.loaders.StartUrlsLoader': 500},
+        'SPIDERFEEDER_INPUT_URI': f'{scheme}input_file.txt'}
+    )
 
     crawler.signals.send_catch_log(signals.spider_opened, spider=crawler.spider)
 
-    assert crawler.spider.start_urls == ['https://url1.com', 'https://url2.com']
+    assert list(crawler.spider.start_urls) == ['https://url1.com', 'https://url2.com']
     mock.assert_called_with(f'{scheme}input_file.txt', crawler.settings)
     assert crawler.stats.get_value(f'spider_feeder/{crawler.spider.name}/url_count') == 2
 
 
 def test_should_override_store(get_crawler, mocker):
     crawler = get_crawler({
+        'EXTENSIONS': {'spider_feeder.loaders.StartUrlsLoader': 500},
         'SPIDERFEEDER_INPUT_URI': 's3://input_file.txt',
         'SPIDERFEEDER_STORES': {
             's3': 'tests.test_loaders.CustomS3Store'
         }
     })
-    StartUrlsLoader.from_crawler(crawler)
 
     crawler.signals.send_catch_log(signals.spider_opened, spider=crawler.spider)
 
     CustomS3Store.assert_called_once_with('s3://input_file.txt', crawler.settings)
 
 
-def test_should_custom_store(get_crawler, mocker):
+def test_should_load_custom_store(get_crawler, mocker):
     crawler = get_crawler({
+        'EXTENSIONS': {'spider_feeder.loaders.StartUrlsLoader': 500},
         'SPIDERFEEDER_INPUT_URI': 'abc://input_file.txt',
         'SPIDERFEEDER_STORES': {
             'abc': 'tests.test_loaders.CustomAbcStore'
         }
     })
-    StartUrlsLoader.from_crawler(crawler)
 
     crawler.signals.send_catch_log(signals.spider_opened, spider=crawler.spider)
 
@@ -84,11 +82,34 @@ def test_uri_format_spider_attributes(get_crawler, mocker):
     mock = mocker.patch('spider_feeder.store.file_store.FileStore')
     mock().__iter__.return_value = iter([('https://url1.com', {}), ('https://url2.com', {})])
 
-    crawler = get_crawler({'SPIDERFEEDER_INPUT_URI': '%(dir)s/%(input_file)s.txt'})
+    crawler = get_crawler({
+        'EXTENSIONS': {'spider_feeder.loaders.StartUrlsLoader': 500},
+        'SPIDERFEEDER_INPUT_URI': '%(dir)s/%(input_file)s.txt'}
+    )
     crawler.spider.dir = '/tmp'
     crawler.spider.input_file = 'spider_input'
-    StartUrlsLoader.from_crawler(crawler)
 
     crawler.signals.send_catch_log(signals.spider_opened, spider=crawler.spider)
 
     mock.assert_called_with('/tmp/spider_input.txt', crawler.settings)
+
+
+def test_load_start_requests(get_crawler, mocker):
+    mock = mocker.patch('spider_feeder.store.file_store.FileStore')
+    store_data = [
+        ('https://url1.com', {'url_id': '1'}),
+        ('https://url2.com', {'url_id': '2'}),
+        ('https://url2.com', {'url_id': '2'}),
+    ]
+    mock().__iter__.return_value = iter(store_data)
+
+    crawler = get_crawler({
+        'EXTENSIONS': {'spider_feeder.loaders.StartRequestsLoader': 500},
+        'SPIDERFEEDER_INPUT_URI': 'input_file.csv',
+    })
+
+    crawler.signals.send_catch_log(signals.spider_opened, spider=crawler.spider)
+
+    request_data = [(r.url, r.meta) for r in crawler.spider.start_requests()]
+    assert request_data == store_data
+    assert crawler.stats.get_value(f'spider_feeder/{crawler.spider.name}/url_count') == 3